qZPfQhxL: implement AbstractBeanDefinition
